resolved dependency issues, app can be run now

+++ src/main/java/com/sda/eventine/repository/EventRepository.java
@@ -3,5 +3,5 @@ package com.sda.eventine.repository;
 import com.sda.eventine.model.Event;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventRepository extends CrudRepository<Event, Integer> {
+public interface EventRepository extends CrudRepository<Event, Long> {
 }
